+++ core/Lucy/Search/Searcher.c
@@ -51,7 +51,7 @@ Hits*
 Searcher_Hits_IMP(Searcher *self, Obj *query, uint32_t offset,
                   uint32_t num_wanted, SortSpec *sort_spec) {
     Query   *real_query = Searcher_Glean_Query(self, query);
-    uint32_t doc_max    = Searcher_Doc_Max(self);
+    uint32_t doc_max    = (uint32_t)Searcher_Doc_Max(self);
     uint32_t wanted     = offset + num_wanted > doc_max
                           ? doc_max
                           : offset + num_wanted;
